CPPOST-94 Convert to VC14: Fix CxxTest
https://issues.shibboleth.net/jira/browse/CPPOST-94

Somewhere in the move of the vc10 project I removed a target
directory in the perl scripts (I actually had removed them all
but I thought I'd put them back).  This fix makes sure that all
the generated files in samlTest are in Projects\vc??\samltest
(which is where the obj files go as well).