SSPCPP-684 Use new function in ReloadableXMLFile to work round XERCES Bug
https://issues.shibboleth.net/jira/browse/SSPCPP-684

Rather than writing the DOM out to the spool file and then commtting it,
we exploit's ReloadableXMLFile new ability to write to the spool file
as the data is coming over the wire.

The commit is as before, and we add a few deletes of the spool file in
prior to exiting with an exception.